chore: cleanup function delcaration types
