+++ include/lora.h
@@ -170,11 +170,11 @@ void selectreceiver();
 void unselectreceiver();
 void writeReg(byte addr, byte value);
 byte readReg(byte addr);
-static void opmode (uint8_t mode);
+void opmode (uint8_t mode);
 static void opmodeLora();
 void SetupLoRa();
 bool receive(char *payload);
 void receivepacket();
-static void configPower (int8_t pw);
+void configPower (int8_t pw);
 static void writeBuf(byte addr, byte *value, byte len);
 void txlora(byte *frame, byte datalen);
