+++ src/lora.c
@@ -30,7 +30,7 @@ char message[256];
 
 byte receivedbytes;
 
-byte hello[32] = "HELLO";
+
 
 void die(const char *s)
 {
@@ -73,7 +73,7 @@ void writeReg(byte addr, byte value)
     unselectreceiver();
 }
 
-static void opmode (uint8_t mode) {
+void opmode (uint8_t mode) {
     writeReg(REG_OPMODE, (readReg(REG_OPMODE) & ~OPMODE_MASK) | mode);
 }
 
@@ -226,7 +226,7 @@ void receivepacket() {
     } // dio0=1
 }
 
-static void configPower (int8_t pw) {
+void configPower (int8_t pw) {
     if (sx1272 == false) {
         // no boost used for now
         if(pw >= 17) {
@@ -281,55 +281,4 @@ void txlora(byte *frame, byte datalen) {
     opmode(OPMODE_TX);
 
     printf("send: %s\n", frame);
-}
-
-int main (int argc, char *argv[]) {
-
-    if (argc < 2) {
-        printf ("Usage: argv[0] sender|rec [message]\n");
-        exit(1);
-    }
-
-    wiringPiSetup () ;
-    pinMode(ssPin, OUTPUT);
-    pinMode(dio0, INPUT);
-    pinMode(RST, OUTPUT);
-
-    wiringPiSPISetup(CHANNEL, 500000);
-
-    SetupLoRa();
-
-    if (!strcmp("sender", argv[1])) {
-        // enter standby mode (required for FIFO loading))
-        opmode(OPMODE_STANDBY);
-
-        writeReg(RegPaRamp, (readReg(RegPaRamp) & 0xF0) | 0x08); // set PA ramp-up time 50 uSec
-
-        configPower(23);
-
-        printf("Send packets at SF%i on %.6lf Mhz.\n", sf,(double)freq/1000000);
-        printf("------------------\n");
-
-        if (argc > 2)
-            strncpy((char *)hello, argv[2], sizeof(hello));
-
-        while(1) {
-            txlora(hello, strlen((char *)hello));
-            delay(5000);
-        }
-    } else {
-
-        // radio init
-        opmode(OPMODE_STANDBY);
-        opmode(OPMODE_RX);
-        printf("Listening at SF%i on %.6lf Mhz.\n", sf,(double)freq/1000000);
-        printf("------------------\n");
-        while(1) {
-            receivepacket(); 
-            delay(1);
-        }
-
-    }
-
-    return (0);
 }
